==> 팀 프로젝트 - 사자감성/CODE/02_corona_excel_to_csv.ipynb <==
### 코로나 데이터
* 코로나 데이터 처리
    * 기간 : 2020-01-20 ~ 2021-08-31
* 데이터 프레임 생성
    * 전체 데이터 프레임                    -> corona_df
    * 주말 제거된 데이터 프레임(토, 일 제거)    -> corona_coin_df
* 각각의 csv 파일 생성
    * 전체 데이터           -> 01_corona.csv        -> 코로나 vs 뉴스 상관관계 분석에 사용
    * 주말 제거된 데이터      -> 02_corona_without_weekend   -> 코로나 vs 주식, 코인 상관관계 분석에 사용

import pandas as pd
from datetime import datetime

xlsx = pd.read_excel("./EXCEL/코로나바이러스감염증-19_확진환자_발생현황_210928.xlsx")
xlsx.head(6)

xlsx.shape

xlsx.isnull().sum()

xlsx.loc[3]

col = xlsx.loc[3]
columns = []
for i in col :
    columns.append(i)
    print(i)

xlsx.columns = columns
xlsx.head(6)

# 상단 3개 행 제거
xlsx_tmp = xlsx.drop([0, 1, 2, 3, 4])

xlsx_tmp.reset_index(drop = True, inplace = True)
xlsx_tmp.head(3)

xlsx_tmp.info()

### 코로나 확진자 수 데이터 타입 변경

corona = xlsx_tmp[["일자", "계(명)"]]
corona["계(명)"] = corona["계(명)"].astype("int")
corona.head(3)

corona.info()

### 코로나 날짜 데이터 타입 변경

corona["일자"] = pd.to_datetime(corona["일자"], unit = "ns")
# corona["일자"].apply(lambda _ : datetime.strptime(_, "%Y-%m-%d %H:%M:%S"))
# corona_df["일자"] = corona_df["일자"].apply(lambda _ : datetime.strptime(_, "%Y-%m-%d %H:%M:%S"))

corona.info()

### 날짜가 2021년 9월 1일 이전인 데이터만 추출
* 2020년 1월 20일 ~ 2021년 8월 31일 데이터

corona = corona.loc[corona["일자"] < '2021-09-01', :]
corona.info()

corona.tail()

### 주말 제거된 데이터 생성

corona["일자"][6].strftime("%A")

corona["일자"].dt.strftime("%A")

corona["일자"].dt.weekday # 0 : Monday, 6 : Sunday

# 5(토요일) : 84, 6(일요일) : 84
# 590 - 84 - 84 = 422
corona_kospi = corona.loc[(corona["일자"].dt.weekday != 5) & (corona["일자"].dt.weekday != 6), :]
corona_kospi.shape

corona_kospi.reset_index(drop = True, inplace = True)
corona_kospi.tail()

### 컬럼명 수정

corona_kospi.columns = ["날짜", "코로나 확진자"]
corona.columns = ["날짜", "코로나 확진자"]

corona_kospi.info()

### DataFrame

print("corona shape :", corona.shape)
print("corona_kospi shape : ", corona_kospi.shape)

corona.to_csv("./CSV/01_corona.csv", index = False)
corona_kospi.to_csv("./CSV/02_corona_without_weekend.csv", index = False)
