==> 팀 프로젝트 - 사자감성/CODE/03_news_data.ipynb <==
### 뉴스 데이터 수집 및 처리, csv 파일 생성

* 빅카인즈 공공데이터 홈페이지에서 뉴스 데이터 크롤링
* 로그인 후 코로나 키워드로 뉴스 검색
* 이틀씩 날짜 지정 후 크롤링

from selenium import webdriver
from selenium.webdriver.common.keys import Keys
import time
import datetime

# 일자 list 생성
start = datetime.datetime.strptime("2020-01-20", "%Y-%m-%d")
end = datetime.datetime.strptime("2021-09-01", "%Y-%m-%d")
date_tmp = [start + datetime.timedelta(days = x) for x in range(0, (end-start).days)]
date_list = []

for date in date_tmp :
    # print(date.strftime("%Y%m%d"))
    # print(date)
    date_list.append(date.strftime("%Y%m%d"))
# print(len(date_list))


# 다운로드 파일 폴더 지정 - option 지정
directory_location = r'/Users/seungwoomun/Documents/Github/LikeLion_13th_DataCourse/03. 팀 프로젝트/EXCEL'
options = webdriver.ChromeOptions()
prefs = {'download.default_directory' : directory_location}
options.add_experimental_option('prefs', prefs)

url = "https://www.bigkinds.or.kr/v2/news/index.do"
driver = webdriver.Chrome("./DATA/chromedriver", options = options)
driver.get(url)

# implicitly wait
# 10초로 설정하면 페이지가 로딩되는데 10초까지 기다린다.
# 만약 페이지 로딩이 2초에 완료되었다면 더 기다리지 않고 다음 코드 실행한다.
driver.implicitly_wait(10)

# xpath
# login_button : //*[@id="header"]/div[1]/div/div[2]/button[1]
# email_field : //*[@id="login-user-id"]
# password_field : //*[@id="login-user-password"]
# login_button2 : //*[@id="login-btn"]
# search_field : //*[@id="total-search-key"]
# corona_word : //*[@id="suggests-wrap"]/li/button
# date_tap : //*[@id="collapse-step-1-body"]/div[3]/div/div[1]/div[1]/a
# start_date : //*[@id="search-begin-date"]
# end_date : //*[@id="search-end-date"]
# search_button : //*[@id="search-foot-div"]/div[2]/button[2]
# result_button : //*[@id="collapse-step-3"]
# download_button : //*[@id="analytics-data-download"]/div[3]/button
# news_tap : //*[@id="collapse-step-1"]

# 로그인 버튼 선택
login_button = driver.find_element_by_xpath('//*[@id="header"]/div[1]/div/div[2]/button[1]')
login_button.click()
time.sleep(1)

# 이메일, 비밀번호 입력 후 로그인 버튼 선택
email_field = driver.find_element_by_xpath('//*[@id="login-user-id"]')
password_field = driver.find_element_by_xpath('//*[@id="login-user-password"]')
login_button2 = driver.find_element_by_xpath('//*[@id="login-btn"]')
time.sleep(1)

email_field.send_keys("city1616@naver.com")
password_field.send_keys("Seungwoo!32")
login_button2.click()
time.sleep(1)

# 뉴스 검색 입력창에 코로나 입력
search_field = driver.find_element_by_xpath('//*[@id="total-search-key"]')
search_field.send_keys("((코로나19) OR (코로나) OR (코로나 바이러스) OR (신종 코로나바이러스) OR (COVID-19) OR (코비드19))")
search_field.click()

# corona_word = driver.find_element_by_xpath('//*[@id="suggests-wrap"]/li/button')
# corona_word.click()
# time.sleep(1)

# 기간 탭 선택 후 가져올 기간 설정
date_tap = driver.find_element_by_xpath('//*[@id="collapse-step-1-body"]/div[3]/div/div[1]/div[1]/a')
date_tap.click()

start_date = driver.find_element_by_xpath('//*[@id="search-begin-date"]')
# start_date.clear()
start_date.send_keys(Keys.BACKSPACE * 10)
time.sleep(1)
start_date.send_keys("20200120")

end_date = driver.find_element_by_xpath('//*[@id="search-end-date"]')
# end_date.clear()
end_date.send_keys(Keys.BACKSPACE * 10)
time.sleep(1)
end_date.send_keys("20200121")

search_button = driver.find_element_by_xpath('//*[@id="search-foot-div"]/div[2]/button[2]')
search_button.click()
driver.implicitly_wait(10)

# 제일 하단으로 스크롤
driver.execute_script("window.scrollTo(0, document.body.scrollHeight);")

result_button = driver.find_element_by_xpath('//*[@id="collapse-step-3"]')
result_button.click()
driver.implicitly_wait(10)

download_button = driver.find_element_by_xpath('//*[@id="analytics-data-download"]/div[3]/button')
download_button.click()
driver.implicitly_wait(10)
time.sleep(1)

driver.execute_script("window.scrollTo(0, 0)")
news_tap = driver.find_element_by_xpath('//*[@id="collapse-step-1"]/span')
# ------ for 시작 -------
for i in range(268, 590, 2) :
    news_tap.click()

    start_date.send_keys(Keys.BACKSPACE * 10)
    time.sleep(1)
    start_date.send_keys(date_list[i])

    end_date.send_keys(Keys.BACKSPACE * 10)
    time.sleep(1)
    end_date.send_keys(date_list[i + 1])
    
    search_button.click()
    driver.implicitly_wait(10)
    
    # 제일 하단으로 스크롤
    driver.execute_script("window.scrollTo(0, document.body.scrollHeight);")
    
    result_button.click()
    driver.implicitly_wait(10)
    
    download_button.click()
    driver.implicitly_wait(10)
    time.sleep(15)

    driver.execute_script("window.scrollTo(0, 0)")

date_list[268]
