==> 팀 프로젝트 - 사자감성/CODE/04_news_data_to_csv.ipynb <==
### 뉴스 데이터 xlsx to csv
* 엑셀 파일의 뉴스 데이터를 모두 가져와 csv 파일로 저장
    * news_all_data.csv : 뉴스 데이터 전체
    * news_data_five_columns.csv : 뉴스 데이터 5개 컬럼
    * news_title.csv : 날짜, 뉴스 제목 데이터

import pandas as pd
import datetime

# 일자 list 생성
start = datetime.datetime.strptime("2020-01-20", "%Y-%m-%d")
end = datetime.datetime.strptime("2021-09-01", "%Y-%m-%d")
date_tmp = [start + datetime.timedelta(days = x) for x in range(0, (end-start).days)]
date_list = []

for date in date_tmp :
    date_list.append(date.strftime("%Y%m%d"))

# excel file name : NewsResult_20200120-20200121
# skiprows = [0]
# 1. 706
# 2. 1191
# 3. 848
news_tmp = pd.read_excel("./EXCEL/news data/NewsResult_20200120-20200121.xlsx")
pd.set_option("display.max_rows", 100)

len(date_list)

print("news_tmp shape :", news_tmp.shape)

col = news_tmp.columns
print("col len :", len(col))
for i in col :
    print(i, end = ", ")

news_data = news_tmp.iloc[1:, :]
news_data.columns = col
print("news_data shape :", news_data.shape)

# 2 ~ len(date_list)
# result2 = pd.concat([df1,df2], ignore_index=True)
# 705 + 1190 + 847 = 2742

for i in range(2, len(date_list), 2) :
    news_tmp = pd.read_excel(f"./EXCEL/news data/NewsResult_{date_list[i]}-{date_list[i + 1]}.xlsx", skiprows = [0])
    news_tmp.columns = col
    news_data = pd.concat([news_data, news_tmp], ignore_index = True)
    print(int(i / 2), "data shape :", news_tmp.shape)
    #     print("data length :", len(news_tmp))
    
# print(type(news_data))
print("data shape :", news_data.shape)

len(news_data)

news_data.columns

news_data.columns = ['뉴스 식별자', '날짜', '언론사', '기고자', '제목', '통합 분류1', '통합 분류2', '통합 분류3',
       '사건/사고 분류1', '사건/사고 분류2', '사건/사고 분류3', '인물', '위치', '기관', '키워드',
       '특성추출(가중치순 상위 50개)', '본문', 'URL', '분석제외 여부']

res_col = ["날짜", "제목", "키워드", "특성추출(가중치순 상위 50개)", "본문"]
news_data_five_columns = news_data[res_col].copy()
news_data_five_columns.head(3)

title_col = ["날짜", "제목"]
news_title_df = news_data[title_col].copy()
news_title_df.head(3)

news_data_five_columns.info()

news_title_df.info()

### 날짜 데이터 타입 변경

news_data["날짜"] = news_data["날짜"].astype("str")
news_data["날짜"] = pd.to_datetime(news_data["날짜"], format = "%Y%m%d")
# news_data["일자"] = pd.to_datetime(news_data["일자"], unit = "ns")
# news_data["일자"] = news_data["일자"].apply(lambda _ : datetime.strftime(_, "%Y%m%d"))
news_data.head(3)

news_data_five_columns["날짜"] = news_data_five_columns["날짜"].astype("str")
news_data_five_columns["날짜"] = pd.to_datetime(news_data_five_columns["날짜"], format = "%Y%m%d")
# news_data["일자"] = pd.to_datetime(news_data["일자"], unit = "ns")
# news_data["일자"] = news_data["일자"].apply(lambda _ : datetime.strftime(_, "%Y%m%d"))
news_data_five_columns.head(3)

news_title_df["날짜"] = news_title_df["날짜"].astype("str")
news_title_df["날짜"] = pd.to_datetime(news_title_df["날짜"], format = "%Y%m%d")
# news_data["일자"] = pd.to_datetime(news_data["일자"], unit = "ns")
# news_data["일자"] = news_data["일자"].apply(lambda _ : datetime.strftime(_, "%Y%m%d"))
news_title_df.head(3)

print(news_data.info())
print("=" * 50)
print(news_data_five_columns.info())
print("=" * 50)
print(news_title_df.info())

print("news_data shape :", news_data.shape)
print("news data five columns shape :", news_data_five_columns.shape)
print("news title df shape :", news_title_df.shape)

### 최종 데이터 프레임
* news_data : 전체 데이터
* news_data_five_columns : 날짜, 제목, 키워드, 특성추출, 본문 컬럼만 가지고 있는 데이터
* news_title_df : 날짜, 제목 컬럼만 가지고 있는 데이터

news_data.to_csv("./CSV/03_news_all_data.csv", index = False)
news_data_five_columns.to_csv("./CSV/04_news_data_five_columns.csv", index = False)
news_title_df.to_csv("./CSV/05_news_title.csv", index = False)
