==> 팀 프로젝트 - 사자감성/CODE/05_news_cnt_columns.ipynb <==
import pandas as pd

news_data = pd.read_csv("./CSV/05_news_title.csv")

news_data.head(3)

news_data.info()

### 날짜 데이터 타입 변경

news_data["날짜"] = news_data["날짜"].astype("str")
news_data["날짜"] = pd.to_datetime(news_data["날짜"], format = "%Y-%m-%d")
# news_data["일자"] = pd.to_datetime(news_data["일자"], unit = "ns")
# news_data["일자"] = news_data["일자"].apply(lambda _ : datetime.strftime(_, "%Y%m%d"))
news_data.head(3)

news_data.info()

news_data.groupby(news_data["날짜"]).count()

news_cnt = news_data.groupby(news_data["날짜"]).count()
# news_df["일자"] = news_df.index
news_cnt.reset_index(inplace = True)
news_cnt.columns = ["날짜", "뉴스 개수"]
# type(news_df)
news_cnt.head(3)
# news_df.columns
# news_df.info()

news_cnt.info()

news_cnt.to_csv("./CSV/06_news_cnt.csv", index = False)
