==> 팀 프로젝트 - 사자감성/CODE/06_bitcoin.ipynb <==
### 비트코인 1일 종가 데이터
* 코로나 기간 : 2020-01-20 ~ 2021-08-31
* 3년 데이터 기간 : 2018-01-01 ~ 2021-08-31
* 데이터 프레임 생성
    * 코로나 기간 : bitcoin
    * 3년 데이터 기간 : bitcoin_3_year
* 각각의 csv 파일 생성
    * 코로나 기간 : 07_bitcoin.csv -> 코로나 확진자 수 vs 비트코인 상관관계 분석을 위해 생성
    * 3년 데이터 기간 : 08_bitcoin_3_year.csv -> 코로나 전후 비교를 위해 생성

import pandas as pd
import pybithumb
import pandas as pd

btc_tmp = pybithumb.get_ohlcv("BTC")

# print(type(btc))
# close = btc[‘close’]
# print(close)

print("btc type :", type(btc_tmp))
print(btc_tmp)

btc_tmp.columns

bitcoin = pd.DataFrame({"비트코인" : btc_tmp.loc["2020-01-20":"2021-08-31"]["close"]})
bitcoin.head(3)

bitcoin.reset_index(inplace = True)
bitcoin.columns = ["날짜", "비트코인"]
bitcoin.head(3)

bitcoin_3_year = pd.DataFrame({"비트코인" : btc_tmp.loc["2018-01-01":"2021-08-31"]["close"]})
bitcoin_3_year.head(3)

bitcoin_3_year.reset_index(inplace = True)
bitcoin_3_year.columns = ["날짜", "비트코인"]
bitcoin_3_year.head(3)

print("bitcoin shape :", bitcoin.shape)
print("bitcoin_3_year shape : ", bitcoin_3_year.shape)

bitcoin.to_csv("./CSV/07_bitcoin.csv", index = False)
bitcoin_3_year.to_csv("./CSV/08_bitcoin_3_year.csv", index = False)
