==> 팀 프로젝트 - 사자감성/CODE/07_kospi.ipynb <==
### 코스피 1일 종가 데이터
* 코로나 기간 : 2020-01-20 ~ 2021-08-31
* 3년 데이터 기간 : 2018-01-01 ~ 2021-08-31
* 데이터 프레임 생성
    * 코로나 기간 : kospi
    * 3년 데이터 기간 : kospi_3_year
* 각각의 csv 파일 생성
    * 코로나 기간 : 09_kospi.csv -> 코로나 확진자 수 vs 코스피 상관관계 분석을 위해 생성
    * 3년 데이터 기간 : 10_kospi_3_year.csv -> 코로나 전후 비교를 위해 생성

import pandas as pd
import yfinance as yf

df = yf.download("^KS11", start = "2020-01-20", end = "2021-08-30")
print(df.info())

df.index

# df.to_csv("kospi.csv")

kospi = pd.DataFrame({"코스피" : df["Close"]})
kospi.info()

kospi.reset_index(inplace = True)
kospi.columns = ["날짜", "코스피"]
kospi.tail(3)

df_tmp = yf.download("^KS11", start = "2018-01-01", end = "2021-08-31")
print(df_tmp.info())

kospi_3_year = pd.DataFrame({"코스피" : df_tmp["Close"]})
kospi_3_year.info()

kospi_3_year.reset_index(inplace = True)
kospi_3_year.columns = ["날짜", "코스피"]
kospi_3_year.tail(3)

print(kospi.info())
print(kospi_3_year.info())

print("kospi shape :", kospi.shape)
print("kospi_3_year shape : ", kospi_3_year.shape)

kospi.to_csv("./CSV/09_kospi.csv", index = False)
kospi_3_year.to_csv("./CSV/10_kospi_3_year.csv", index = False)

# import matplotlib.pyplot as plt
 
# msft = yf.Ticker("066570.KS")
# #msft.info
# df =msft.history(start ='2018-01-01', end = '2020-03-04')
# df =msft.history(period = '1y')
# df =msft.history(period = 'max')
# plt.plot(list(df['Open']))
