==> 팀 프로젝트 - 사자감성/CODE/08_data_merge.ipynb <==
### 데이터 병합
* 코인 데이터 : bit.csv
* 코로나 데이터
    * 전체 : corona.csv
    * 주말 제거 : corona_coin.csv
* 주식 데이터 : kospi.csv
* 뉴스 데이터
    * 전체 : news_all_data.csv
    * 5개 컬럼 : news_data_five_columns.csv
        * 날짜, 제목, 키워드, 특성추출(가중치순 상위 50개), 본문
    * 2개 컬럼 : news_title.csv
        * 날짜, 제목
    * 날짜별 뉴스 개수 : news_cnt.csv
        * 날짜, 뉴스 개수
* 결과 데이터
    * 코로나(전체), 주식, 코인, 뉴스 병합된 csv

import pandas as pd
from datetime import datetime

corona = pd.read_csv("./CSV/01_corona.csv")
corona_without_weekend = pd.read_csv("./CSV/02_corona_without_weekend.csv")
bitcoin = pd.read_csv("./CSV/07_bitcoin.csv")
kospi = pd.read_csv("./CSV/09_kospi.csv")
news_cnt = pd.read_csv("./CSV/06_news_cnt.csv")

print("corona shape :", corona.shape)
print("corona_without_weekend shape :", corona_without_weekend.shape)
print("bitcoin shape :", bitcoin.shape)
print("kospi shape :", kospi.shape)
print("news_cnt shape :", news_cnt.shape)

print(corona.info())
print("=" * 50)
print(corona_without_weekend.info())
print("=" * 50)
print(bitcoin.info())
print("=" * 50)
print(kospi.info())
print("=" * 50)
print(news_cnt.info())

### 날짜 데이터 타입 변경

corona["날짜"] = pd.to_datetime(corona["날짜"], format = "%Y-%m-%d")
corona_without_weekend["날짜"] = pd.to_datetime(corona_without_weekend["날짜"], format = "%Y-%m-%d")
bitcoin["날짜"] = pd.to_datetime(bitcoin["날짜"], format = "%Y-%m-%d")
kospi["날짜"] = pd.to_datetime(kospi["날짜"], format = "%Y-%m-%d")
news_cnt["날짜"] = pd.to_datetime(news_cnt["날짜"], format = "%Y-%m-%d")

print(corona.info())
print("=" * 50)
print(corona_without_weekend.info())
print("=" * 50)
print(bitcoin.info())
print("=" * 50)
print(kospi.info())
print("=" * 50)
print(news_cnt.info())

### corona, bitcoin merge

# result_df = corona.merge(bitcoin, left_on = "일자", right_on = "일자")
data_merge_df = pd.merge(corona, bitcoin, on = "날짜")
data_merge_df.head(3)

data_merge_df.info()

### data_merge_df, kospi merge

# result_df = pd.merge(corona, bitcoin, on = "일자")
data_merge_df = data_merge_df.merge(kospi, left_on = "날짜", right_on = "날짜", how = "outer")
data_merge_df.head(3)

data_merge_df.info()

data_merge_df.head(20)

### data_merge_df, news_df merge

# result_df["news_cnt"] = news_data["일자"].value_counts(sort = False)
# result_df
data_merge_df = data_merge_df.merge(news_cnt, on = "날짜")
data_merge_df.head(20)

data_merge_df.tail(20)

### CSV 파일로 저장
* index 제거

data_merge_df.to_csv("./CSV/11_data_merge.csv", index = False)
