==> 팀 프로젝트 - 사자감성/CODE/11_news_title_wordcloud.ipynb <==
### 날짜, 제목 컬럼을 가지고 있는 뉴스 데이터 불러오기

import pandas as pd
from wordcloud import WordCloud, STOPWORDS
from konlpy.tag import Okt
from konlpy.tag import Hannanum
import time

news_data = pd.read_csv("./CSV/05_news_title.csv")
news_data.info()

news_data.head(3)

len(news_data)

start = time.time() # 코드 시작 시간

title = ""
for idx, val in enumerate(news_data["제목"]) :
    # print(idx, val)
    # title.append(val)
    title = title + val + " "
    if idx % 10000 == 0 :
        print(idx, val)
    
print("title length :", len(title))
print("소요 시간 :", time.time() - start) # 현재시간 - 시작시간 = 실행 시간

title[59534000 : ]

news_data.tail(3)

t = Okt()
doc_nouns = t.nouns(doc_ko)
